Fix the hex formatting, use 64-bit values

+++ src/StatusDisplay/StatusDisplayCell.cc
@@ -149,8 +149,9 @@ namespace BUTool{
     to the given buffer. The buffer will be modified in place.
     */
     
-    // Read the 32-bit value from the register
-    uint32_t val = regIO->ReadRegister(address);
+    // Read the 32-bit value from the register and convert to 64-bit value
+    // so that it's ready for printing
+    uint64_t val = regIO->ReadRegister(address);
 
     // Now, do the formatting
     std::string fmtString = "%";
